+++ src/Exceptions/UnsupportedColumnTypeException.php
@@ -16,7 +16,7 @@ class UnsupportedColumnTypeException extends \Exception
     public static function forColumnKeyVal(string $columnKey, mixed $columnVal): static
     {
         $msg = \sprintf('Unsupported column type (%s): %s', \get_debug_type($columnVal), $columnKey);
-        return self($msg);
+        return new self($msg);
     }
 // empty
 }