+++ CHANGES.md
@@ -16,6 +16,12 @@ Fast and flexible PHP library for text tables.
 
 ## Changes
 
+* dev
+  * Added `Separator` row class, and `Table::addSeparator()` helper method.
+  * Fixed `DuplicateColumnException` and `UnsupportedColumnAlignException` exceptions
+  * Improved documentation.
+
+
 * 2.6.0 (2023-10-12)
   * Added `CompactRenderer`.
   * Renderers can now control rendering of top/bottom/separator rows.